fix: functional

+++ openpolvobackend/migrations/000012_add_smtp_email_chat_skip_confirmation.up.sql
@@ -1,1 +1,1 @@
-ALTER TABLE laele_user_smtp_settings ADD COLUMN email_chat_skip_confirmation INTEGER NOT NULL DEFAULT 0;
+ALTER TABLE laele_user_smtp_settings ADD COLUMN email_chat_skip_confirmation TINYINT(1) NOT NULL DEFAULT 0;
